+++ starters/react/resources/js/types/global.d.ts
@@ -1,7 +1,8 @@
-import type { SharedProps } from '@/types/shared';
+import type { FlashData, SharedProps } from '@/types/shared';
 
 declare module '@inertiajs/core' {
     export interface InertiaConfig {
         sharedPageProps: SharedProps;
+        flashDataType: FlashData;
     }
 }